Commit #1.1: Found a quick bug
When inputting a decimal number less than 1 and has a leading zero (example: 0.5), parsing it as an integer truncates the input to 0, causing the program to think that there is an invalid input. Switched to parsing as a float.

+++ src/Vform.js
@@ -95,7 +95,7 @@ class Vform extends Component {
 		}
 
 		// checking for inputs with value zero
-		else if (parseInt(this.state.s1) === 0 || parseInt(this.state.s2) === 0 || parseInt(this.state.s3) === 0) {
+		else if (parseFloat(this.state.s1) === 0 || parseFloat(this.state.s2) === 0 || parseFloat(this.state.s3) === 0) {
 			return true;
 		}
